feat: change tracking script from leadfeeder to leadinfo
